requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-amazon==3.7.0
pydantic==1.10.13
typing-extensions==4.5.0
pyarrow==12.0.1
pandas==2.1.0
boto3==1.31.0
